+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/data/CrbKenyaMetropolRequestData.java
@@ -18,7 +18,6 @@
  */
 package org.apache.fineract.portfolio.loanaccount.data;
 
-import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -28,7 +27,7 @@ public class CrbKenyaMetropolRequestData {
     private String application_ref_no;
     private String identity_number;
     private String identity_type;
-    private BigDecimal loan_amount;
+    private Integer loan_amount;
     private Integer report_reason;
     private Integer id;
     private Integer clientId;
@@ -62,7 +61,7 @@ public class CrbKenyaMetropolRequestData {
     }
 
     public CrbKenyaMetropolRequestData(Integer report_type, String application_ref_no, String identity_number, String identity_type,
-            BigDecimal loan_amount, Integer report_reason) {
+            Integer loan_amount, Integer report_reason) {
         this.report_type = report_type;
         this.application_ref_no = application_ref_no;
         this.identity_number = identity_number;
